+++ lib/deimos/backends/kafka.rb
@@ -6,6 +6,12 @@ module Deimos
     class Kafka < Deimos::PublishBackend
       include Phobos::Producer
 
+      # Shut down the producer if necessary.
+      def self.shutdown_producer
+        producer.sync_producer_shutdown if producer.respond_to?(:sync_producer_shutdown)
+        producer.kafka_client&.close
+      end
+
       # :nodoc:
       def self.execute(producer_class:, messages:)
         Deimos.instrument(